move calcFftOfSamples to utils folder
